# api/handlers/workspace.go
package handlers

import (
	"net/http"

	ws_manager "github.com/EO-DataHub/eodhp-workspace-manager/models"
	"github.com/EO-DataHub/eodhp-workspace-services/api/services"
	"github.com/gorilla/mux"
	_ "github.com/lib/pq"
)

// Trick compiler to keep import for swag annotation
var _ = ws_manager.WorkspaceSettings{}

// CreateWorkspace handles HTTP requests for creating a new workspace.
func CreateWorkspace(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		// Get a token from keycloak so we can interact with it's API
		if !ensureKeycloakToken(w, svc.KC) {
			return
		}

		svc.CreateWorkspaceService(w, r)
	}
}

// @Summary Get a list of workspaces you are a member of
// @Description Retrieve a list of workspaces for the authenticated user.
// @Tags Workspace Management
// @Accept json
// @Produce json
// @Success 200 {array} ws_manager.WorkspaceSettings
// @Failure 401 {object} string
// @Failure 500 {object} string
// @Router /workspaces [get]
func GetWorkspaces(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		// Get a token from keycloak so we can interact with it's API
		if !ensureKeycloakToken(w, svc.KC) {
			return
		}

		svc.GetWorkspacesService(w, r)
	}
}

// @Summary Get a workspace by ID
// @Description Retrieve a specific workspace using its ID for the authenticated user.
// @Tags Workspace Management
// @Accept json
// @Produce json
// @Param workspace-id path string true "Workspace ID" // Workspace ID from the URL path
// @Success 200 {object} ws_manager.WorkspaceSettings
// @Failure 400 {object} string
// @Failure 401 {object} string
// @Failure 500 {object} string
// @Router /workspaces/{workspace-id} [get]
func GetWorkspace(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		// Get a token from keycloak so we can interact with it's API
		if !ensureKeycloakToken(w, svc.KC) {
			return
		}

		svc.GetWorkspaceService(w, r)
	}
}

// DeleteWorkspace handles HTTP requests for deleting a workspace
func DeleteWorkspace(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		// Get a token from keycloak so we can interact with it's API
		if !ensureKeycloakToken(w, svc.KC) {
			return
		}

		svc.DeleteWorkspaceService(w, r)
	}
}

// UpdateWorkspace handles HTTP requests for updating a specific workspace by ID.
// This is a placeholder for the actual implementation.
func UpdateWorkspace(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		vars := mux.Vars(r)
		workspaceID := vars["workspace-id"]

		// placeholder for the implementation
		http.Error(w, "Failed to update workspace "+workspaceID, http.StatusInternalServerError)
	}
}

// PatchWorkspace handles HTTP requests for partially updating a specific workspace by ID.
// This is a placeholder for the actual implementation.
func PatchWorkspace(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		vars := mux.Vars(r)
		workspaceID := vars["workspace-id"]

		// placeholder for the implementation
		http.Error(w, "Failed to patch workspace "+workspaceID, http.StatusInternalServerError)
	}
}

// @Summary Get users of a workspace
// @Description Retrieve a list of users who are members of the specified workspace.
// @Tags Workspace Management
// @Accept json
// @Produce json
// @Param workspace-id path string true "Workspace ID"
// @Success 200 {array} models.User
// @Failure 400 {object} string
// @Failure 401 {object} string
// @Failure 500 {object} string
// @Router /workspaces/{workspace-id}/users [get]
func GetUsers(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		// Get a token from keycloak so we can interact with it's API
		if !ensureKeycloakToken(w, svc.KC) {
			return
		}

		svc.GetUsersService(w, r)
	}
}

// @Summary Get a user of a workspace
// @Description Retrieve details of a specific user that is a member of the specified workspace.
// @Tags Workspace Management
// @Accept json
// @Produce json
// @Param workspace-id path string true "Workspace ID"
// @Param username path string true "Username"
// @Success 200 {object} models.User
// @Failure 400 {object} string
// @Failure 401 {object} string
// @Failure 404 {object} string
// @Failure 500 {object} string
// @Router /workspaces/{workspace-id}/users/{username} [get]
func GetUser(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		// Get a token from keycloak so we can interact with it's API
		if !ensureKeycloakToken(w, svc.KC) {
			return
		}

		svc.GetUserService(w, r)
	}
}

// @Summary Add a user to a workspace
// @Description Add a user to the specified workspace by providing the workspace ID and username.
// @Tags Workspace Management
// @Accept json
// @Produce json
// @Param workspace-id path string true "Workspace ID"
// @Param username path string true "Username"
// @Success 204 {string} string
// @Failure 400 {object} string
// @Failure 401 {object} string
// @Failure 404 {object} string
// @Failure 500 {object} string
// @Router /workspaces/{workspace-id}/users/{username} [put]
func AddUser(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		// Get a token from keycloak so we can interact with it's API
		if !ensureKeycloakToken(w, svc.KC) {
			return
		}

		svc.AddUserService(w, r)
	}
}

// @Summary Remove a user from a workspace
// @Description Remove a user from the specified workspace by providing the workspace ID and username.
// @Tags Workspace Management
// @Accept json
// @Produce json
// @Param workspace-id path string true "Workspace ID"
// @Param username path string true "Username"
// @Success 204 {string} string
// @Failure 400 {object} string
// @Failure 401 {object} string
// @Failure 404 {object} string
// @Failure 500 {object} string
// @Router /workspaces/{workspace-id}/users/{username} [delete]
func RemoveUser(svc *services.WorkspaceService) http.HandlerFunc {

	return func(w http.ResponseWriter, r *http.Request) {

		// Get a token from keycloak so we can interact with it's API
		if !ensureKeycloakToken(w, svc.KC) {
			return
		}

		svc.RemoveUserService(w, r)
	}
}