# docs/openapi.yaml
openapi: 3.0.0
info:
  description: This is the API for the EODHP Workspace Services.
  title: EODHP Workspace Services API
  contact: {}
  version: v1
paths:
  /accounts:
    get:
      description: Retrieve a list of billing accounts owned by the authenticated user.
      tags:
        - Billing and Billing Accounts
      summary: Get billing accounts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/models.Account"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
  "/accounts/{id}":
    get:
      description: Retrieve details of a specific billing account by its unique ID.
      tags:
        - Billing and Billing Accounts
      summary: Get a billing account
      parameters:
        - description: Account ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Account"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
  /workspaces:
    get:
      description: Retrieve a list of workspaces for the authenticated user.
      tags:
        - Workspace Management
      summary: Get a list of workspaces you are a member of
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/models.WorkspaceSettings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
  "/workspaces/{workspace-id}":
    get:
      description: Retrieve a specific workspace using its ID for the authenticated user.
      tags:
        - Workspace Management
      summary: Get a workspace by ID
      parameters:
        - description: Workspace ID
          name: workspace-id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.WorkspaceSettings"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
...
  "/workspaces/{workspace-id}/users":
    get:
      description: Retrieve a list of users who are members of the specified workspace.
      tags:
        - Workspace Management
      summary: Get users of a workspace
      parameters:
        - description: Workspace ID
          name: workspace-id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/models.User"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
  "/workspaces/{workspace-id}/users/{username}":
    get:
      description: Retrieve details of a specific user that is a member of the
        specified workspace.
      tags:
        - Workspace Management
      summary: Get a user of a workspace
      parameters:
        - description: Workspace ID
          name: workspace-id
          in: path
          required: true
          schema:
            type: string
        - description: Username
          name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.User"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
    put:
      description: Add a user to the specified workspace by providing the workspace ID
        and username.
      tags:
        - Workspace Management
      summary: Add a user to a workspace
      parameters:
        - description: Workspace ID
          name: workspace-id
          in: path
          required: true
          schema:
            type: string
        - description: Username
          name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
    delete:
      description: Remove a user from the specified workspace by providing the
        workspace ID and username.
      tags:
        - Workspace Management
      summary: Remove a user from a workspace
      parameters:
        - description: Workspace ID
          name: workspace-id
          in: path
          required: true
          schema:
            type: string
        - description: Username
          name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
...
components:
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    models.Account:
      type: object
      properties:
        accountOpeningReason:
          type: string
        accountOwner:
          type: string
        billingAddress:
          type: string
        createdAt:
          type: string
        id:
          type: string
        name:
          type: string
        organizationName:
          type: string
        status:
          type: string
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/models.WorkspaceSettings"
    models.BlockStore:
      type: object
      properties:
        access_point_id:
          type: string
        mount_point:
          type: string
        name:
          type: string
        store_id:
          type: string
    models.ObjectStore:
      type: object
      properties:
        access_point_arn:
          type: string
        access_url:
          type: string
        bucket:
          type: string
        env_var:
          type: string
        host:
          type: string
        name:
          type: string
        prefix:
          type: string
        store_id:
          type: string
    models.Stores:
      type: object
      properties:
        block:
          type: array
          items:
            $ref: "#/components/schemas/models.BlockStore"
        object:
          type: array
          items:
            $ref: "#/components/schemas/models.ObjectStore"
    models.User:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        id:
          type: string
        lastName:
          type: string
        username:
          type: string
    models.WorkspaceSettings:
      type: object
      properties:
        account:
          type: string
        id:
          type: string
        last_updated:
          type: string
        name:
          type: string
        owner:
          type: string
        status:
          type: string
        stores:
          type: array
          items:
            $ref: "#/components/schemas/models.Stores"
...